TYP: move ``vectorize`` stubs to ``lib._function_base_impl``
